agentic_data_analysis/agentic_data_analysis/requirements.txt
pandas
numpy
scipy
scikit-learn
plotly
nltk
google-adk
kaleido